requirements/dev-requirements.txt
pytest==8.0.2
ruff
black
isort
flask-testing
locust
pytest-cov